+++ backend/src/main/java/com/jamesmcdonald/backend/login/AccountResponse.java
@@ -1,9 +1,11 @@
 package com.jamesmcdonald.backend.login;
 
+import java.math.BigDecimal;
+
 public record AccountResponse(
          Long id,
          String cardNumber,
-         int balance,
+         BigDecimal balance,
          String name,
          String email,
          String phone
